+++ caffe2/core/tensor.h
@@ -611,8 +611,6 @@ CAFFE2_API void ReinitializeAndCopyFrom(
     const Tensor& src,
     bool async = false);
 
-CAFFE_DECLARE_PREALLOCATED_KNOWN_TYPE(12, Tensor)
-
 using TensorCPU = Tensor;
 
 constexpr int k_limit_default_ = 1000;
